requirements.txt
streamlit==1.39.0
anthropic==0.40.0
asyncio
aiohttp==3.10.11
python-dotenv==1.0.1
reportlab==4.2.5
markdown==3.7
pydantic==2.10.3
tenacity==9.0.0
plotly==5.24.1
pandas==2.2.3
numpy==2.2.1